Add AoC 2022 Day 16 Part 2 input and (correct) solution
This reverts commits all the way back to the "probably correct, but far,
far too slow!" (and adds the "toExplore.clear()" to reduce memory usage).
Turns out it is indeed correct, but takes about 25 minutes on my UM690,
and about 15GB of RAM!

Will see if we can optimise it a bit more.  Using hashmaps instead of
maps might be a good idea.