AoC22D16P2: optimise - use unordered_map; normalise states
This brings the runtime on UM690 to 684.77 and roughly halves the memory
usage.

Now that I've completed it, I've looked at the solutions - people are
generally using techniques which occurred to me, but struck me as either
incorrect or not very useful - shows how much I know :p Although some
people's logic *is* pretty ropey, and I don't trust their solutions to
be correct outside of the given Puzzle Input(s).

Lots of sub-second(!) runtimes!

Will definitely re-visit this; a DFS is probably better than my BFS, as
it allows early pruning before the state-space gets too huge i.e. with a
DFS, we actually get *a* solution very quickly (unlike with BFS) and can
use that to abort other solutions rapidly.