AoC23D18P1: begin to optimise to deal with very large numToDig
The eventual, optimised solution I come up with for this will ultimately
form  the basis of the Part 1 solution, but it's easier to incubate it
in Part 1's code as I'll be able to check the optimised solution against
Part 1's original "brute force" implementation.

For now, just fork the brute force implementation out into its own
method, and add a means of storing the input data in sensible data
structures that can be used by the brute force implementation and the
upcoming optimised version (and also use int64_t instead of ints).