Add (helper for) AoC 2023 Day 20 Part 2 (correct) solution
I just solved this "manually" using the code below to help me (turns out
that we have qz, cq, jx, tt, as inputs to qn, which is a conjunction
which is the sole input to rx; the four inputs to qn each fire a High
pulse cyclically).

The next commit will make this less manual, though it will still rely on
the specialist structure of my (and everybody elses, presumably) input.