Add AoC 2023 Day X Part 2 helper
Feels like I cheated on this one as I was struggling to come up with a
workable general solution, but when examining the puzzle input, I
noticed an anomaly: two of the hailstones had identical startY and
velocityY, which enables us to (manually) deduce everything else quite
easily :/ That's the approach I've taken here.

Not sure what a general solution would look like, but note that each
hailstone allows us to add three more equations with just one extra
unknown (on top of the six initial unknowns: the smashing hailstone's
starting coordinate and its velocity), so maybe that's a path? If we
consider just three hailstones, that's nine unknowns over nine
equations, so likely solvable. Definitely can't be bothered, if so:
sounds like an algebraic-manipulation nightmare XD

Another approach might be to assume that the absolute velocity of the
smashing hailstone is quite small (as it turns out it is) and
brute-force all components of it (assuming, say, -500 -> +500) - this
would probably work, too, though still feels a bit cheat-y.

Yet another approach might be to try and solve the 2D case (just X&Y
collisions) - this will probably be much less burdensome in terms of
linear algebra, and (I suspect) almost immediately extend to the 3D
case.