Add AoC 2018 Day 21 Part 1 input and helper
Another "reverse-engineering" one, though I did minimal
reverse-engineering and just looked for the "escape" condition; found
the first occurrence of it, and set register 0 to the value necessary to
get it to terminate.

Don't actually know what the program does XD

The solution doesn't really do much except verify that giving register 0
the value 3173684 causes it to terminate.