Add AoC 2018 Day 21 Part 2 helper
Again, this is a reverse-engineering problem, so the helper doesn't
really do anything.

It was instrumental in speeding up evaluation of the program: a large
part of the runtime was spent laboriously right-shifting register 1 by 8
bits, and that chunk has been replaced by simply directly performing
this right-shift in C++ (see "Shortcut.").

After that, we just have to look for the circularity in the final
comparison of register 3 values with register 0, and pick the value just
before the first repetition of these register 3 values.