Add AoC 2016 Day 11 Part 2 (correct) solution
Needs substantial cleanup and documentation, but not in the mood right
now :p

Basic idea: exploit the symmetries to create a "normalised"
representation.  Two states with the same normalised representation
should take the same number of steps to reach the final state.  There
are massively fewer distinct normalised representations of states than
there are states.