Add AoC 2016 Day 25 Part 1 input and (correct) solution
This is a reverse engineering problem; the code will print out the
answer, then dump a load of debugging info out for ever and ever.

As with all AoCs I've done so far, there is no Day 25 Part 2, so this is
it for AoC 2016 - will add a screenshot of the fully-completed calendar
later!